+++ source/packages/@aws-accelerator/accelerator/lib/validate-environment-config.ts
@@ -22,8 +22,6 @@ import { AcceleratorResourcePrefixes } from '../utils/app-utils';
 export interface ValidateEnvironmentConfigProps {
   readonly acceleratorConfigTable: cdk.aws_dynamodb.ITable;
   readonly newOrgAccountsTable: cdk.aws_dynamodb.ITable;
-  readonly newCTAccountsTable: cdk.aws_dynamodb.ITable;
-  readonly controlTowerEnabled: boolean;
   readonly organizationsEnabled: boolean;
   readonly commitId: string;
   readonly stackName: string;
@@ -83,7 +81,7 @@ export class ValidateEnvironmentConfig extends Construct {
       sid: 'dynamodb',
       effect: cdk.aws_iam.Effect.ALLOW,
       actions: ['dynamodb:PutItem'],
-      resources: [props.newOrgAccountsTable.tableArn, props.newCTAccountsTable?.tableArn],
+      resources: [props.newOrgAccountsTable.tableArn],
     });
     const ddbConfigTablePolicy = new cdk.aws_iam.PolicyStatement({
       sid: 'dynamodbConfigTable',
@@ -95,7 +93,7 @@ export class ValidateEnvironmentConfig extends Construct {
       sid: 'kms',
       effect: cdk.aws_iam.Effect.ALLOW,
       actions: ['kms:Encrypt', 'kms:Decrypt', 'kms:GenerateDataKey*', 'kms:DescribeKey'],
-      resources: [props.newOrgAccountsTable.encryptionKey!.keyArn, props.newCTAccountsTable.encryptionKey!.keyArn],
+      resources: [props.newOrgAccountsTable.encryptionKey!.keyArn],
     });
     const cloudformationPolicy = new cdk.aws_iam.PolicyStatement({
       sid: 'cloudformation',
@@ -157,8 +155,6 @@ export class ValidateEnvironmentConfig extends Construct {
       properties: {
         configTableName: props.acceleratorConfigTable.tableName,
         newOrgAccountsTableName: props.newOrgAccountsTable.tableName,
-        newCTAccountsTableName: props.newCTAccountsTable?.tableName || '',
-        controlTowerEnabled: props.controlTowerEnabled,
         organizationsEnabled: props.organizationsEnabled,
         commitId: props.commitId,
         stackName: props.stackName,